requirements.txt
flask==3.0.0
flask-cors==4.0.0
pymongo==4.6.1
pandas==2.1.1
numpy==1.24.3
python-dotenv==1.0.0
